leaderboard service documentation, additional methods, project cleanup
started documenting the LeaderboardService calls, still wip

added the UpdatePlayer methods to the LeaderboardService. These methods will execute updates on a subset of players based on given arguments. These updates are random by nature to give a natural feel to the updates as they are designed to be used in automation.

removing some sprites that were added at the beginning and never used

also removed the `UpdateLeaderboard` call in the `UpdatePlayer` method. I don't want to force updating the leaderboard, I think it's best to leave it under the individual dev's control