adding automated updates
The `LeaderboardUpdateController` class will handle automated updates based on the set update cycle (daily, weekly, monthly). On init the controller will check if the update date has passed, if so it will set the new update date and then execute the leaderboard update. I've written it to be adaptive and not just add (n) for each cycle, if the cycle is monthly and the updates happen on the 15th the system should maintain that cycle of updating on the 15th.

new enum `LeaderboardUpdateCycle` created to support the update system